requirements.txt
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-pn532==2.4.6
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
colorzero==2.0
gpiozero==2.0.1
hx711==1.1.2.3
lgpio==0.2.2.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.15.0
